# src/resources/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Blocos de Notas",
    "version": "1.0.0",
    "description": "API REST para gerenciar blocos de notas com autenticação JWT"
  },
  "paths": {
    "/login": {
      "post": {
        "summary": "Autenticar usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/notes": {
      "post": {
        "summary": "Criar um bloco de notas",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bloco criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos"
          },
          "401": {
            "description": "Token não fornecido ou inválido"
          }
        }
      },
      "get": {
        "summary": "Listar todos os blocos de notas",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de blocos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "notes": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Note" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token não fornecido ou inválido"
          }
        }
      }
    },
    "/notes/{id}": {
      "get": {
        "summary": "Buscar um bloco de notas",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Bloco encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteResponse"
                }
              }
            }
          },
          "404": {
            "description": "Bloco não encontrado"
          },
          "401": {
            "description": "Token não fornecido ou inválido"
          }
        }
      },
      "put": {
        "summary": "Atualizar um bloco de notas",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bloco atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteResponse"
                }
              }
            }
          },
          "404": {
            "description": "Bloco não encontrado"
          },
          "401": {
            "description": "Token não fornecido ou inválido"
          }
        }
      },
      "delete": {
        "summary": "Remover um bloco de notas",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Bloco removido com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Bloco não encontrado"
          },
          "401": {
            "description": "Token não fornecido ou inválido"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "NoteInput": {
        "type": "object",
        "required": ["title", "content"],
        "properties": {
          "title": { "type": "string" },
          "content": { "type": "string" }
        }
      },
      "Note": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "content": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "NoteResponse": {
        "type": "object",
        "properties": {
          "note": {
            "$ref": "#/components/schemas/Note"
          }
        }
      }
    }
  }
}